% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_errors.R
\name{catch_errors}
\alias{catch_errors}
\title{Catch input errors before using lookup tables}
\usage{
catch_errors(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for 'qt-2000' and
'qt-legio' method).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item 'qt' for QuantiTray method with 51 wells.
  \item 'qt-2000' for QuantiTray/2000 method with 97 wells.
  \item 'qt-legio' for QuantiTray/Legiolert method with 96 wells.
}}
}
\value{
Error if any input erro is detected; nothing if not.
}
\description{
This function catches input errors.
This function is and internal function for the package "quantitray".
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
