% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_qt_errors.R
\name{catch_qt_errors}
\alias{catch_qt_errors}
\title{Catch input errors before using lookup tables}
\usage{
catch_qt_errors(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for method
\code{qt-2000} and \code{qt-legio}).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item \code{qt} for QuantiTray method with 51 wells.
  \item \code{qt-2000} for QuantiTray/2000 method with 97 wells.
  \item \code{qt-legio} for QuantiTray/Legiolert method with 96 wells.
}}
}
\value{
Error if any input erro is detected; nothing if not.
}
\description{
This function catches input errors for the \code{quantify_} family of
functions.
This function is an internal function for the package \pkg{quantitray}.
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
