% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lookuptable.R
\name{load_lookuptable}
\alias{load_lookuptable}
\title{Load data from QuantiTray lookup tables}
\usage{
load_lookuptable(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for 'qt-2000' and
'qt-legio' method).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item 'qt' for QuantiTray method with 51 wells.
  \item 'qt-2000' for QuantiTray/2000 method with 97 wells.
  \item 'qt-legio' for QuantiTray/Legiolert method with 96 wells.
}}
}
\value{
Tibble of numeric data values from lookup table.
}
\description{
This function loads lookup tables for the quantify_ family of functions.
This function is and internal function for the package "quantitray".
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
