% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpn.R
\docType{package}
\name{mpn}
\alias{mpn}
\alias{mpn-package}
\title{mpn: Convert QuantiTray results to MPN.}
\description{
An R package to calculate MPN (most probable number) from IDEXX Quanti-Tray,
Quanti-Tray/2000, and Quanti-Tray/Legiolert tests
}
\section{Quick reference}{

tbc
}

\section{Full documentation}{

Full documentation for this package is available at
http://www.janknappe.com/mpn/
}

