% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanti_51.R
\docType{data}
\name{quanti_51}
\alias{quanti_51}
\title{Conversion table from Quanti-Tray readings to MPN}
\format{A tibble with 52 rows and 4 variables}
\source{
\href{https://idexxcom-live-b02da1e51e754c9cb292133b-9c56c33.aldryn-media.com/filer_public/0f/45/0f4542b3-2b85-432b-aa79-e480ce5ad99f/qt51mpntable.pdf}{IDEXX Technologies Ltd}
}
\usage{
data(quanti_51)
}
\description{
Results from an IDEXX Quanti-Tray are expressed in counts of positive
wells. This is a lookup table to convert these counts into MPN (most
probable number) with upper and lower 95\% confidence limits expressed
in MPN/100 ml. The data set includes the following columns:
}
\details{
\itemize{
  \item count_large. number of positive wells (0 -- 51)
  \item MPN_95lo. lower 95\% confidence interval (0 -- 146.1)
  \item MPN_mean. most probable number (0 -- 200.5)
  \item MPN_95hi. upper 95\% confidence interval (3.67 -- 387.6)
}
}
\examples{
data(quanti_51)

}
\references{
IDEXX Technologies, Quanti-Tray MPN Table.
(\href{https://www.idexx.com/en/water/resources/mpn-generator/}{IDEXX Technologies Ltd})
}
\keyword{datasets}
