% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_95lo.R
\name{quantify_95lo}
\alias{quantify_95lo}
\title{Convert IDEXX QuantiTray results to lower 95 \% confidence interval}
\usage{
quantify_95lo(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for 'qt-2000' method).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item 'qt' for QuantiTray method with 51 wells.
  \item 'qt-2000' for QuantiTray/2000 method with 97 wells.
}}
}
\value{
Numerical value: lower 95 \% confidence interval of MPN/100ml in
  sample.
}
\description{
This function converts counts of positive wells (large and small) into lower
95 \% confidence interval values in units of MPN/100ml for IDEXX
QuantiTray and QuantiTray/2000 trays.
}
\examples{
quantify_95lo(large = 42, method = "qt")
quantify_95lo(large = 42, small = 42, method = "qt-2000")

}
\references{
https://www.idexx.com/en/water/resources/mpn-generator/
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
