% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_all.R
\name{quantify_all}
\alias{quantify_all}
\title{Convert IDEXX QuantiTray results to MPN with confidence intervals}
\usage{
quantify_all(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for 'qt-2000' method).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item 'qt' for QuantiTray method with 51 wells.
  \item 'qt-2000' for QuantiTray/2000 method with 97 wells.
}}
}
\value{
List of numerical values:
  \itemize{
    \item 'MPN_lo' as lower 95\% confidence interval.
    \item 'MPN' as point estimate of MPN/100ml in sample.
    \item 'MPN_hi' as upper 95\% confidence interval.
  }
}
\description{
This function converts counts of positive wells (large and small) into point
estimete MPN (most probable number) values plus lower and upper 95\%
confidence intervals values in units of MPN/100ml for IDEXX QuantiTray and
QuantiTray/2000.
}
\examples{
quantify_all(large = 42, method = "qt")
quantify_all(large = 42, small = 42, method = "qt-2000")

}
\references{
https://www.idexx.com/en/water/resources/mpn-generator/
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
