% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_mpn.R
\name{quantify_mpn}
\alias{quantify_mpn}
\title{Convert IDEXX QuantiTray results to MPN}
\usage{
quantify_mpn(large, small, method)
}
\arguments{
\item{large}{Number of positve large wells (required for all methods).}

\item{small}{Number of positve small wells (required for method
\code{qt-2000} and \code{qt-legio}).}

\item{method}{Defines which QuantiTray method was used:
\itemize{
  \item \code{qt} for QuantiTray method with 51 wells.
  \item \code{qt-2000} for QuantiTray/2000 method with 97 wells.
  \item \code{qt-legio} for QuantiTray/Legiolert method with 96 wells.
}}
}
\value{
Numerical value: Point estimate of MPN/100ml in sample.
}
\description{
This function converts counts of positive wells (large and small) into point
estimete MPN (most probable number) values in units of MPN/100ml for IDEXX
QuantiTray, QuantiTray/2000, and QuantiTray/Legiolert trays.
}
\examples{
quantify_mpn(large = 42, method = "qt")
quantify_mpn(large = 42, small = 42, method = "qt-2000")
quantify_mpn(large = 5, small = 42, method = "qt-legio")

}
\references{
\url{https://www.idexx.com/en/water/resources/mpn-generator/}
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
