% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitray.R
\docType{package}
\name{quantitray}
\alias{quantitray}
\alias{quantitray-package}
\title{quantitray: Convert Quanti-Tray results to MPN.}
\description{
An R package to calculate MPN (most probable number) from IDEXX Quanti-Tray,
Quanti-Tray/2000, and Quanti-Tray/Legiolert tests
}
\section{Quick reference}{

A suite of functions to convert Quanti-Tray results into MPN (most probable
number) readings.
  \itemize{
    \item \code{quantify_mpn()} outputs the point estimate for MPN.
    \item \code{quantify_95lo()} outputs the lower 95\% confidence interval value.
    \item \code{quantify_95hi()} outputs the upper 95\% confidence interval value.
  }
}

\section{See also}{

  \itemize{
    \item Full documentation for this package is available at
        \url{https://www.janknappe.com/quantitray}
    \item Report bugs at \url{https:github.com/jknappe/quantitray/issues}
   }
}

