% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitray.R
\docType{package}
\name{quantitray}
\alias{quantitray}
\alias{quantitray-package}
\title{quantitray: Convert Quanti-Tray results to MPN.}
\description{
An R package to calculate MPN (most probable number) from IDEXX Quanti-Tray,
Quanti-Tray/2000, and Quanti-Tray/Legiolert tests
}
\section{Quick reference}{

A suite of functions to convert Quanti-Tray results into MPN (most probable
number) readings.
  \itemize{
    \item 'quantify_mpn' outputs the point estimate for MPN.
    \item 'quantify_95lo' outputs the lower 95\% confidence interval value.
    \item 'quantify_95hi' outputs the upper 95\% confidence interval value.
    \item 'quantify_all' outputs all the above as a list..
  }
}

\section{Full documentation}{

Full documentation for this package is available at
https://www.janknappe.com/quantitray
}

